Добавляет новые эндпоинты, сущности, подключение и работу с БД H2

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -32,25 +32,25 @@ public class UserController {
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    public User addFriend(@PathVariable("id") Long userId,
-                          @PathVariable("friendId") Long friendId) {
+    public User addFriend(@PathVariable("id") Integer userId,
+                          @PathVariable("friendId") Integer friendId) {
         return userService.addFriend(userId, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    public User removeFriend(@PathVariable("id") Long userId,
-                             @PathVariable("friendId") Long friendId) {
+    public User removeFriend(@PathVariable("id") Integer userId,
+                             @PathVariable("friendId") Integer friendId) {
         return userService.removeFriend(userId, friendId);
     }
 
     @GetMapping("/{id}/friends")
-    public Collection<User> getFriends(@PathVariable("id") Long userId) {
+    public Collection<User> getFriends(@PathVariable("id") Integer userId) {
         return userService.getFriends(userId);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public Collection<User> getCommonFriends(@PathVariable("id") Long userId,
-                                             @PathVariable("otherId") Long otherUserId) {
+    public Collection<User> getCommonFriends(@PathVariable("id") Integer userId,
+                                             @PathVariable("otherId") Integer otherUserId) {
         return userService.getCommonFriends(userId, otherUserId);
     }
 }
